+++ src/app/collective-rewards/rewards/backers/index.ts
@@ -1,17 +1,16 @@
 export * from './hooks'
 export * from './context'
+export * from './ABIBackers'
 import { Rewards } from './Rewards'
 import { RewardsCard } from './RewardsCard'
 import { AllTimeShare } from './AllTimeShare'
 import { ClaimableRewards } from './ClaimableRewards'
 import { BackerRewardsTable } from './BackerRewardsTable'
-import { ABI } from './ABI'
 
 export {
   Rewards as BackerRewards,
   RewardsCard as BackerRewardsCard,
   AllTimeShare as BackerAllTimeShare,
   ClaimableRewards as BackerClaimableRewards,
-  ABI as BackerABI,
   BackerRewardsTable,
 }
