+++ src/app/collective-rewards/shared/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useGaugesGetFunction'
 export * from './useReadBackersManager'
 export * from './useGetCyclePayout'
 export * from './useGetEstimatedBackersRewardsPct'
+export * from './useGetABI'
